+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/example/SearchFieldWithDefaultLabel.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import SearchField from 'terra-search-field';
+
+const SearchFieldWithLabel = () => (
+  <SearchField isLabelVisible />
+);
+
+export default SearchFieldWithLabel;