+++ packages/terra-paginator/src/ProgressivePaginator.jsx
@@ -31,6 +31,14 @@ const propTypes = {
    */
   itemCountPerPage: PropTypes.number,
   /**
+   * Allows user to set the heading id of the page.
+   */
+  ariaLabelledBy: PropTypes.string,
+  /**
+   * Allows user to set the custom paginator label.
+   */
+  ariaLabel: PropTypes.string,
+  /**
    * @private
    * The intl object to be injected for translations.
    */
@@ -82,76 +90,101 @@ class ProgressivePaginator extends React.Component {
       intl,
       totalCount,
       itemCountPerPage,
+      ariaLabelledBy,
+      ariaLabel,
       pageLabel,
     } = this.props;
     const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const { selectedPage } = this.state;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
+    const paginatorAriaLabel = ariaLabel || 'pagination';
 
     const { messageId, messageAttributes } = getPageLabel(pageLabel, selectedPage, totalPages);
 
-    return (
-      <div className={cx('paginator', 'progressive', theme.className)} role="navigation" aria-label="pagination">
+    const fullViewChildren = (
+      <>
         <div>
           {
             intl.formatMessage({ id: messageId }, messageAttributes)
           }
         </div>
-        <div>
+        <ul className={cx('progressive-list')}>
           {
             totalCount && (
-              <PaginatorButton
-                ariaDisabled={selectedPage === 1}
-                ariaLabel={intl.formatMessage({ id: 'Terra.paginator.first' })}
-                className={cx(['nav-link', selectedPage === 1 ? 'is-disabled' : null])}
-                disabled={selectedPage === 1}
-                onClick={this.handlePageChange(1)}
-                onKeyDown={this.handlePageChange(1)}
-              >
-                {intl.formatMessage({ id: 'Terra.paginator.first' })}
-              </PaginatorButton>
+              <li>
+                <PaginatorButton
+                  ariaDisabled={selectedPage === 1}
+                  ariaLabel={intl.formatMessage({ id: 'Terra.paginator.first' })}
+                  className={cx(['nav-link', selectedPage === 1 ? 'is-disabled' : null])}
+                  disabled={selectedPage === 1}
+                  onClick={this.handlePageChange(1)}
+                  onKeyDown={this.handlePageChange(1)}
+                >
+                  {intl.formatMessage({ id: 'Terra.paginator.first' })}
+                </PaginatorButton>
+              </li>
             )
           }
-          <PaginatorButton
-            ariaDisabled={selectedPage === 1}
-            ariaLabel={intl.formatMessage({ id: 'Terra.paginator.previous' })}
-            className={cx(['nav-link', 'previous', selectedPage === 1 ? 'is-disabled' : null])}
-            disabled={selectedPage === 1}
-            onClick={this.handlePageChange(previousPageIndex)}
-            onKeyDown={this.handlePageChange(previousPageIndex)}
-          >
-            <span className={cx('icon')} />
-            {intl.formatMessage({ id: 'Terra.paginator.previous' })}
-          </PaginatorButton>
-          <PaginatorButton
-            ariaDisabled={selectedPage === totalPages}
-            ariaLabel={intl.formatMessage({ id: 'Terra.paginator.next' })}
-            className={cx(['nav-link', 'next', selectedPage === totalPages ? 'is-disabled' : null])}
-            disabled={selectedPage === totalPages}
-            onClick={this.handlePageChange(nextPageIndex)}
-            onKeyDown={this.handlePageChange(nextPageIndex)}
-          >
-            {intl.formatMessage({ id: 'Terra.paginator.next' })}
-            <span className={cx('icon')} />
-          </PaginatorButton>
+          <li>
+            <PaginatorButton
+              ariaDisabled={selectedPage === 1}
+              ariaLabel={intl.formatMessage({ id: 'Terra.paginator.previous' })}
+              className={cx(['nav-link', 'previous', selectedPage === 1 ? 'is-disabled' : null])}
+              disabled={selectedPage === 1}
+              onClick={this.handlePageChange(previousPageIndex)}
+              onKeyDown={this.handlePageChange(previousPageIndex)}
+            >
+              <span className={cx('icon')} />
+              {intl.formatMessage({ id: 'Terra.paginator.previous' })}
+            </PaginatorButton>
+          </li>
+          <li>
+            <PaginatorButton
+              ariaDisabled={selectedPage === totalPages}
+              ariaLabel={intl.formatMessage({ id: 'Terra.paginator.next' })}
+              className={cx(['nav-link', 'next', selectedPage === totalPages ? 'is-disabled' : null])}
+              disabled={selectedPage === totalPages}
+              onClick={this.handlePageChange(nextPageIndex)}
+              onKeyDown={this.handlePageChange(nextPageIndex)}
+            >
+              {intl.formatMessage({ id: 'Terra.paginator.next' })}
+              <span className={cx('icon')} />
+            </PaginatorButton>
+          </li>
           {
             (totalCount) && (
-              <PaginatorButton
-                ariaDisabled={selectedPage === totalPages}
-                ariaLabel={intl.formatMessage({ id: 'Terra.paginator.last' })}
-                className={cx(['nav-link', selectedPage === totalPages ? 'is-disabled' : null])}
-                disabled={selectedPage === totalPages}
-                onClick={this.handlePageChange(totalPages)}
-                onKeyDown={this.handlePageChange(totalPages)}
-              >
-                {intl.formatMessage({ id: 'Terra.paginator.last' })}
-              </PaginatorButton>
+              <li>
+                <PaginatorButton
+                  ariaDisabled={selectedPage === totalPages}
+                  ariaLabel={intl.formatMessage({ id: 'Terra.paginator.last' })}
+                  className={cx(['nav-link', selectedPage === totalPages ? 'is-disabled' : null])}
+                  disabled={selectedPage === totalPages}
+                  onClick={this.handlePageChange(totalPages)}
+                  onKeyDown={this.handlePageChange(totalPages)}
+                >
+                  {intl.formatMessage({ id: 'Terra.paginator.last' })}
+                </PaginatorButton>
+              </li>
             )
           }
-        </div>
-      </div>
+        </ul>
+      </>
     );
+
+    const fullView = ariaLabelledBy
+      ? (
+        <nav className={cx('paginator', 'progressive', theme.className)} aria-labelledby={ariaLabelledBy}>
+          {fullViewChildren}
+        </nav>
+      )
+      : (
+        <nav className={cx('paginator', 'progressive', theme.className)} aria-label={paginatorAriaLabel}>
+          {fullViewChildren}
+        </nav>
+      );
+
+    return fullView;
   }
 
   reducedProgressivePaginator() {
@@ -160,17 +193,20 @@ class ProgressivePaginator extends React.Component {
       intl,
       totalCount,
       itemCountPerPage,
+      ariaLabelledBy,
+      ariaLabel,
       pageLabel,
     } = this.props;
     const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const { selectedPage } = this.state;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
     const nextPageIndex = selectedPage === totalPages ? totalPages : selectedPage + 1;
+    const paginatorAriaLabel = ariaLabel || 'pagination';
 
     const { messageId, messageAttributes } = getPageLabel(pageLabel, selectedPage, totalPages);
 
-    return (
-      <div className={cx('paginator', theme.className)} role="navigation" aria-label="pagination">
+    const reducedViewChildren = (
+      <>
         <div>
           {
             (totalCount) && (
@@ -230,8 +266,22 @@ class ProgressivePaginator extends React.Component {
             )
           }
         </div>
-      </div>
+      </>
     );
+
+    const reducedView = ariaLabelledBy
+      ? (
+        <nav className={cx('paginator', theme.className)} aria-labelledby={ariaLabelledBy}>
+          {reducedViewChildren}
+        </nav>
+      )
+      : (
+        <nav className={cx('paginator', theme.className)} aria-label={paginatorAriaLabel}>
+          {reducedViewChildren}
+        </nav>
+      );
+
+    return reducedView;
   }
 
   render() {