+++ packages/terra-button/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 3.68.0 - (August 11, 2023)
+
+* Changed
+  * Reverted `3.66` changes.
+
+## 3.67.0 - (August 11, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.66.0 - (August 8, 2023)
 
 * Changed