+++ packages/terra-form-radio/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 4.39.0 - (August 11, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.38.0 - (August 11, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.37.0 - (August 8, 2023)
 
 * Changed